# Fix to improve performance of ChangesTreeView