--- a/src/GitHub.Api/UI/TreeBuilder.cs
+++ b/src/GitHub.Api/UI/TreeBuilder.cs
@@ -6,7 +6,7 @@ namespace GitHub.Unity
 {
     static class TreeBuilder
     {
-        internal static void BuildChildNode(FileTreeNode parent, FileTreeNode node, List<string> foldedTreeEntries1, Action<FileTreeNode> stateChangeCallback1)
+        internal static void BuildChildNode(FileTreeNode parent, FileTreeNode node, HashSet<string> foldedTreeSet, Action<FileTreeNode> stateChangeCallback)
         {
             if (String.IsNullOrEmpty(node.Label))
             {
@@ -15,7 +15,7 @@ namespace GitHub.Unity
             }
 
             node.RepositoryPath = parent.RepositoryPath.ToNPath().Combine(node.Label);
-            parent.Open = !foldedTreeEntries1.Contains(parent.RepositoryPath);
+            parent.Open = !foldedTreeSet.Contains(parent.RepositoryPath);
 
             // Is this node inside a folder?
             var nodePath = node.Label.ToNPath();
@@ -33,7 +33,7 @@ namespace GitHub.Unity
                     if (child.Label.Equals(root))
                     {
                         found = true;
-                        BuildChildNode(child, node, foldedTreeEntries1, stateChangeCallback1);
+                        BuildChildNode(child, node, foldedTreeSet, stateChangeCallback);
                         break;
                     }
                 }
@@ -42,7 +42,7 @@ namespace GitHub.Unity
                 if (!found)
                 {
                     var p = parent.RepositoryPath.ToNPath().Combine(root);
-                    BuildChildNode(parent.Add(new FileTreeNode(root, stateChangeCallback1) { RepositoryPath = p }), node, foldedTreeEntries1, stateChangeCallback1);
+                    BuildChildNode(parent.Add(new FileTreeNode(root, stateChangeCallback) { RepositoryPath = p }), node, foldedTreeSet, stateChangeCallback);
                 }
             }
             else if (nodePath.ExtensionWithDot == ".meta")
@@ -56,7 +56,7 @@ namespace GitHub.Unity
                     if (child.Label.Equals(searchLabel))
                     {
                         found = true;
-                        BuildChildNode(child, node, foldedTreeEntries1, stateChangeCallback1);
+                        BuildChildNode(child, node, foldedTreeSet, stateChangeCallback);
                         break;
                     }
                 }
@@ -106,6 +106,8 @@ namespace GitHub.Unity
                 }
             }
 
+            var foldedTreeSet = new HashSet<string>(foldedTreeEntries);
+
             // Add new stuff
             for (var index = 0; index < newEntries.Count; ++index)
             {
@@ -136,7 +138,7 @@ namespace GitHub.Unity
                     node.Icon = iconLoaderFunc?.Invoke(gitStatusEntry.ProjectPath);
                 }
 
-                BuildChildNode(tree, node, foldedTreeEntries, stateChangeCallback);
+                BuildChildNode(tree, node, foldedTreeSet, stateChangeCallback);
             }
 
             return tree;