Attempted #23 

performance test used pumps two sequences into `TreeBuilder.BuildTreeRoot()`:
1. Simple Test
  a. The first sequence contains 200 `GitStatusEntry` objects.
    1. There are 10 folders, with 10 files and 10 meta files.
  a. The second sequence contains 300 `GitStatusEntry` objects.
    1. Out of the original dataset every other file is retained.
      a. There are 10 folders, with 5 files and 5 meta files.
      a. Half of the original files can be considered "removed".
      a. An additional 10 folders, with 10 files and 10 meta files.
1. Advanced Test
  a. The first sequence contains 800 `GitStatusEntry` objects.
    1. There are 20 folders, with 20 files and 20 meta files.
  a. The second sequence contains 1200 `GitStatusEntry` objects.
    1. Out of the original dataset every other file is retained.
      a. There are 10 folders, with 10 files and 10 meta files.
      a. Half of the original files can be considered "removed".
      a. An additional 20 folders, with 20 files and 20 meta files.

|                       Method |     Mean |    Error |   StdDev |
|----------------------------- |---------|---------|---------|
 |Basic Test - Currently| 70.71 ms | 1.3993 ms | 1.3743 ms |
 |Basic Test - PR|   26.77 ms | 0.6483 ms | 0.6657 ms |
 |Heavy Test - Currently|754.46 ms | 6.6629 ms | 5.9064 ms |
 |Heavy Test - PR| 110.11 ms | 2.1563 ms | 2.2143 ms |

This branch will let you reproduce the performance test framework:
https://github.com/github-for-unity/Unity/tree/fixes/changes-view-performance-proof
Run the PerformanceTest project un-attached in Release mode.
It will take approx 8-9 minutes to complete.

Overall the PR starts at a third of the time of the current code base, and the amount of time it takes to execute grows bu half of the current code base.

Depends on:
- [x] #120 
- [x] #121
- [x] #133