@StanleyGoldman 2017-07-12
https://gist.github.com/StanleyGoldman/c84e31e174a91e1ff28770feaaf8da29

I've been using this python script to generate data to test with, this is an improvement, but I definitely feel the need for more.