@StanleyGoldman 2017-07-31
It's an nice improvement, but it doesn't solve the overall performance issues.